refactor privacy storage to package
